--- code/spikes/publish.001/src/Pkg.Content.Bundle/common.mts
@@ -1,1 +1,0 @@
-export * from '../common/index.mjs';
