--- code/spikes/publish.001/src/Pkg.Content.Bundle/index.mts
@@ -1,2 +1,0 @@
-export { ContentBundle } from './ContentBundle.mjs';
-export { MarkdownFile } from '../Markdown.File/index.mjs';
