+++ code/spikes/publish.001/src/Pkg/Content.Bundle/ContentBundle.TEST.mts
@@ -1,6 +1,6 @@
 import { Text } from 'sys.text';
 
-import { describe, expect, expectError, it, TestFilesystem } from '../test/index.mjs';
+import { describe, expect, expectError, it, TestFilesystem } from '../../test/index.mjs';
 import { ContentBundle } from './index.mjs';
 
 describe('ContentPipeline.README', () => {
