+++ code/spikes/publish.001/src/Pkg/Content.Bundle/ContentBundle.mts
@@ -1,8 +1,8 @@
-import { t, Path } from '../common/index.mjs';
-import { MarkdownFile } from '../Markdown.File/index.mjs';
+import { t, Path } from '../common.mjs';
+import { MarkdownFile } from '../../Markdown.File/index.mjs';
 import { Filesize } from 'sys.fs';
-import { ContentLog } from '../Pkg.Content.Log/ContentLog.mjs';
-import { BundlePaths } from './Paths.mjs';
+import { ContentLog } from '../Content.Log/ContentLog.mjs';
+import { BundlePaths } from '../Paths.mjs';
 
 type Sources = {
   app: t.Fs; //     The compiled bundle of the content rendering "app" (application).
