+++ code/spikes/publish.001/src/Pkg/Content.Bundle/index.mts
@@ -1,0 +1,1 @@
+export { ContentBundle } from './ContentBundle.mjs';
