+++ code/spikes/publish.001/src/Pkg/common.mts
@@ -1,0 +1,1 @@
+export * from '../common.mjs';
