+++ code/spikes/publish.001/src/Pkg/index.mts
@@ -1,0 +1,2 @@
+export * from './Content.Bundle/index.mjs';
+export * from './Content.Log/index.mjs';
