+++ code/spikes/publish.001/src/Pkg/types.mts
@@ -1,0 +1,2 @@
+export * from './Content.Log/types.mjs';
+export * from './Content.Bundle/types.mjs';
